requirements.txt
streamlit==1.32.0
tensorflow==2.19.0
keras==3.0.5
numpy==1.24.3
Pillow==10.2.0 